requirements.txt
google-auth==2.29.0
google-api-python-client==2.125.0
google-cloud-texttospeech==2.16.3
openai~=1.75.0
protobuf~=4.25.8
requests~=2.32.3